+++ x/subspaces/keeper/hooks_test.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/desmos-labs/desmos/v4/x/subspaces/types"
 )